pyproject.toml
[project]
name = "banana-slides"
version = "0.1.0"
description = "Banana Slides – AI native PPT generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.52.0",
    "openai>=1.0.0",
    "pydantic>=2.9.0",
    "pillow>=12.0.0",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.0.1",
    "reportlab>=4.1.0",
    "werkzeug>=3.0.1",
    "markitdown[all]",
    "tenacity>=9.0.0",
    "alembic>=1.13.0",
    "flask-migrate>=4.0.0",
    "pyjwt>=2.8.0",
    "httpx[socks]>=0.28.1",
    # 可编辑 PPT 导出依赖
    "opencv-python-headless>=4.8.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
